Adds NFS CSI plugin integration test (#7)
The Helm chart depends on a few other CSI components. We have rocks for
these, and we'll be using them in the helm chart as well.

The test will also check the Pod probes to make sure that the service is
up, without waiting for Kubernetes to do the check.

The test will also spawn a NFS server, and create an NGINX Pod requiring
a NFS volume.